Add an additional constraint to Generic class.
The Code of a type is a type-level list of lists. We already
expect the Code to be in SingI. This change additionally expects
each of the inner lists to be in SingI. The change should be
harmless in the sense that the constraint should trivially hold
for all instances anyway. However, some generic functions can be
defined more conveniently, because GHC has more information to
use and distribute.