Add type-level metadata.
We use a type-level representation that is not a GADT.

This is simpler to work with, because it means that the kind of the
type-level metadata is always the same.

Safety is retained, because the whole metadata value is of course
statically known.

This does not yet implement TH generation, and it currently requires
GHC 8.