Use a compact representation internally.
WIP.

We use a vector to represent an NP, and an
Int pair to represent an NS.

We provide the old interface via pattern
synonyms.

The change is experimental, but at least for
sums, it should provide faster access by
avoiding the linear chain of constructors.

Also, deriving this representation for large
datatypes should be cheaper, because the
internal types do not take as much space.