Add AllZip, trans_*, coerce_*, toI_*, fromI_* functions.
These functions are for converting between related structures
that do not have common signatures.

The most common application of these seems to be the scenario
where you have a datatype where every component is wrapped in
a common type constructor application. Then it would be nice to
have this type constructor appear as the constructor argument
to SOP. However, the normal conversion will always instantiate
this to I. Then fromI_SOP and toI_SOP can be used to perform
the desired conversion.