Add a HTrans class capturing the trans_* and coerce_* functions.
It remains to be seen how much type inference is possible.
The two type arguments for HTrans are necessary to avoid
requiring a rank-2 polymorphic kind. In principle, trans_*
can transform one use of a structure into another use of
the same structure, but used at a different kind.

Rank-2 kind polymorphism is, afaik, only available in GHC 8
via TypeInType. Having two type arguments for the class
avoids the problem. The Same type family is then used to
restore some type inference by stating that both type arguments
have to be instantiated to the same structure.