Explicitly quantify the kind variable in `shape` and `lengthSList`
GHC 8.10 (HEAD) implements GHC
[proposal 24](https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0024-no-kind-vars.rst),
which requires kind variables to be quantified alongside type
variables in a top-level `forall`. Currently, the kind variables in
the top-level `forall`s of the `shape` and `lengthSList` functions
are not explicitly quantified, which means that `sop-core` fails to
build on GHC HEAD.

Luckily, the fix is backwards compatible, since the earliest version
of GHC that `sop-core` supports now is GHC 8.0, which includes the
ability to explicitly quantify kind variables. Note that this ability
required the use of the `TypeInType` extension prior to GHC 8.6,
which is why I selectively enable it on old GHCs, but after 8.6 it
only requires `PolyKinds`.