Make sop-core forward-compatible with GHC proposal 229
GHC HEAD now implements
[proposal 229](https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0229-whitespace-bang-patterns.rst),
which makes the parser more sensitive to whitespace around `@`
characters in as-patterns. In particular, `@`s in as-patterns are
no longer permitted to have preceding or trailing whitespace, which
causes `sop-core` to fail to compile on HEAD:

```
[7 of 8] Compiling Data.SOP.Dict    ( src/Data/SOP/Dict.hs, interpreted )

src/Data/SOP/Dict.hs:71:1: error: Parse error in pattern: mapAll2
   |
71 | mapAll2 f d @ Dict = (all2 . mapAll (mapAll f) . unAll2) d
   | ^^^^^^^^^^^
```

The fix is simple: remove the extra whitespace, which this patch
accomplishes.